Moved web3 button down
